requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.47
alembic==1.11.0
psycopg2-binary==2.9.7
python-multipart==0.0.6
aiofiles==23.2.1
Pillow==10.1.0
boto3==1.28.0
httpx==0.24.0
